+++ dramatiq/middleware/__init__.py
@@ -1,6 +1,7 @@
 from .age_limit import AgeLimit  # noqa
 from .callbacks import Callbacks  # noqa
 from .middleware import Middleware, MiddlewareError, SkipMessage  # noqa
+from .pipelines import Pipelines  # noqa
 from .prometheus import Prometheus  # noqa
 from .retries import Retries  # noqa
 from .time_limit import TimeLimit, TimeLimitExceeded  # noqa